backend/requirements.txt
# SurBlend Requirements
# For Raspberry Pi 5 with Python 3.11+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic[email]==2.5.3
email-validator==2.1.0

# API Documentation
# Included with FastAPI

# PDF Generation
reportlab==4.0.8
pypdf==3.17.4

# CSV Processing
pandas==2.1.4
openpyxl==3.1.2

# Linear Programming (for blend optimization)
scipy==1.11.4
numpy==1.26.3

# System Monitoring
psutil==5.9.7

# Date/Time
python-dateutil==2.8.2

# Development Tools (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
isort==5.13.2

# Rate Limiting
slowapi==0.1.9

# Background Tasks (optional)
celery==5.3.4
redis==5.0.1

# Logging
python-json-logger==2.0.7